fix: reorganize resources to internal dir, cleanup provider
