+++ provider/internal/argocd/argocd_types.go
@@ -1,4 +1,4 @@
-package provider
+package argocd
 
 import "github.com/pulumi/pulumi/sdk/v3/go/pulumi"
