+++ provider/internal/resources/vpc/vpc.go
@@ -1,4 +1,4 @@
-package provider
+package vpc
 
 import (
 	"fmt"
